Moved stuff around.

+++ src/Rendering/IRenderBuffer.h
@@ -1,7 +1,7 @@
 
 #pragma once
 
-#include <cstdlib>
+#include <cstdint>
 
 namespace mono
 {
@@ -10,7 +10,7 @@ namespace mono
     public:
 
         virtual ~IRenderBuffer() = default;
-        virtual void UpdateData(const void* data, size_t offset, size_t count) = 0;
+        virtual void UpdateData(const void* data, uint32_t offset, uint32_t count) = 0;
         virtual void Use() const = 0;
         virtual uint32_t Id() const = 0;
     };
